+++ packages/moonstone/CHANGELOG.md
@@ -13,6 +13,10 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/Slider` exports `SliderFactory` and `SliderBaseFactory`
 - `moonstone/IncrementSlider` exports `IncrementSliderFactory` and `IncrementSliderBaseFactory`
 
+### Changed
+
+- `moonstone/VirtualList` and `moonstone/VirtualGridList` prop `component` to be replaced by `itemRenderer`
+
 ### Fixed
 
 - `moonstone/ExpandableItem` to be more performant when animating