+++ packages/ui/CHANGELOG.md
@@ -8,6 +8,10 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 - `ui/Transition` property `clipHeight`
 
+### Changed
+
+- `ui/VirtualList` and `ui/VirtualGridList` prop `component` to be replaced by `itemRenderer`
+
 ### Fixed
 
 - `ui/Transition` animation for `clip` for "up", "left", and "right" directions. This includes a DOM addition to the Transition markup.